Reorganize .prt structures into separate files
