+++ OP2Utility.vcxproj
@@ -163,10 +163,12 @@
     <ClCompile Include="src\Streams\FileWriter.cpp" />
     <ClCompile Include="src\Streams\MemoryWriter.cpp" />
     <ClCompile Include="src\Streams\MemoryReader.cpp" />
+    <ClInclude Include="src\Sprites\Animation.h" />
     <ClInclude Include="src\Sprites\ArtReader.h" />
     <ClInclude Include="src\Sprites\ArtFile.h" />
     <ClInclude Include="src\Sprites\ArtWriter.h" />
     <ClInclude Include="src\Sprites\Color.h" />
+    <ClInclude Include="src\Sprites\ImageMeta.h" />
     <ClInclude Include="src\Sprites\ImageType.h" />
     <ClInclude Include="src\Sprites\PaletteHeader.h" />
     <ClInclude Include="src\Sprites\SectionHeader.h" />
